+++ rlclientlib/vw_model/vw_model.h
@@ -18,7 +18,7 @@ namespace reinforcement_learning { namespace model_management {
 
     int update(const model_data& data, bool& model_ready, api_status* status = nullptr) override;
     int choose_rank(uint64_t rnd_seed, const char* features, std::vector<int>& action_ids, std::vector<float>& action_pdf, std::string& model_version, api_status* status = nullptr) override;
-    int request_decision(const std::vector<const char*>& event_ids, const char* features, std::vector<std::vector<size_t>>& actions_ids, std::vector<std::vector<float>>& action_pdfs, std::string& model_version, api_status* status = nullptr) override;
+    int request_decision(const std::vector<const char*>& event_ids, const char* features, std::vector<std::vector<uint32_t>>& actions_ids, std::vector<std::vector<float>>& action_pdfs, std::string& model_version, api_status* status = nullptr) override;
   private:
     using vw_ptr = std::shared_ptr<safe_vw>;
     using pooled_vw = utility::pooled_object_guard<safe_vw, safe_vw_factory>;
